refactor raster operations into different strategies
